+++ install/migrations/update_10.0.x_to_10.1.0/translations.php
@@ -33,7 +33,11 @@
  * ---------------------------------------------------------------------
  */
 
-Config::deleteConfigurationValues('core', [
+/**
+ * @var \Migration $migration
+ */
+
+$migration->removeConfig([
     'translate_dropdowns',
     'translate_kb',
     'translate_reminders',